Update topic modal edit form for editing icons and banners

+++ DisqussTopics/Controllers/PostController.cs
@@ -92,9 +92,6 @@ namespace DisqussTopics.Controllers
                     videoResultURL = videoResult.SecureUrl.ToString();
                 }
 
-
-
-
                 // Map the postViewModel properties to the post model
                 var post = new Post()
                 {
@@ -188,22 +185,6 @@ namespace DisqussTopics.Controllers
 
             if (ModelState.IsValid)
             {
-                // Get the image result
-                var imageResult = await _imageService.AddImageAsync(postViewModel.UploadImage);
-                string? imageResultURL = string.Empty;
-                if (imageResult.SecureUrl != null)
-                {
-                    imageResultURL = imageResult.SecureUrl.ToString();
-                }
-
-                // Get the video result 
-                var videoResult = await _videoService.AddVideoAsync(postViewModel.UploadVideo);
-                string? videoResultURL = string.Empty;
-                if (videoResult.SecureUrl != null)
-                {
-                    videoResultURL = videoResult.SecureUrl.ToString();
-                }
-
                 var post = await _postRepository
                     .GetPostById(id);
 
@@ -217,38 +198,63 @@ namespace DisqussTopics.Controllers
                 SlugHelper helper = new SlugHelper();
                 string slug = helper.GenerateSlug(postViewModel.Post.Title);
 
-                if (post.Image != null && post.Image.Length > 5) 
-                {
-                    // try to delete the old image
-                    try
-                    {
-                        var fileInfo = new FileInfo(post.Image);
-                        var publicId = Path.GetFileNameWithoutExtension(fileInfo.FullName);
-                        await _imageService.DeleteImageAsync(publicId);
-                    }
-                    catch (Exception)
-                    {
 
-                        ModelState.AddModelError("", "Failed to edit image");
-                        return View(postViewModel);
+                // Get the image result
+                var imageResult = await _imageService.AddImageAsync(postViewModel.UploadImage);
+                string? imageResultURL = string.Empty;
+                if (imageResult.SecureUrl != null)
+                {
+                    imageResultURL = imageResult.SecureUrl.ToString();
+
+                    if (post.Image != null && post.Image.Length > 5)
+                    {
+                        // try to delete the old image
+                        try
+                        {
+                            var fileInfo = new FileInfo(post.Image);
+                            var publicId = Path.GetFileNameWithoutExtension(fileInfo.FullName);
+                            await _imageService.DeleteImageAsync(publicId);
+                        }
+                        catch (Exception)
+                        {
+
+                            ModelState.AddModelError("", "Failed to edit image");
+                            return View(postViewModel);
+                        }
                     }
                 }
-
-                if (post.Video != null && post.Video.Length > 5)
+                else
                 {
-                    // try to delete the old video
-                    try
-                    {
-                        var fileInfo = new FileInfo(post.Video);
-                        var publicId = Path.GetFileNameWithoutExtension(fileInfo.FullName);
-                        await _videoService.DeleteVideoAsync(publicId);
-                    }
-                    catch (Exception)
-                    {
+                    imageResultURL = post.Image;
+                }
 
-                        ModelState.AddModelError("", "Failed to edit video");
-                        return View(postViewModel);
+                // Get the video result 
+                var videoResult = await _videoService.AddVideoAsync(postViewModel.UploadVideo);
+                string? videoResultURL = string.Empty;
+                if (videoResult.SecureUrl != null)
+                {
+                    videoResultURL = videoResult.SecureUrl.ToString();
+
+                    if (post.Video != null && post.Video.Length > 5)
+                    {
+                        // try to delete the old video
+                        try
+                        {
+                            var fileInfo = new FileInfo(post.Video);
+                            var publicId = Path.GetFileNameWithoutExtension(fileInfo.FullName);
+                            await _videoService.DeleteVideoAsync(publicId);
+                        }
+                        catch (Exception)
+                        {
+
+                            ModelState.AddModelError("", "Failed to edit video");
+                            return View(postViewModel);
+                        }
                     }
+                }
+                else
+                {
+                    videoResultURL = post.Video;
                 }
 
                 post.Title = postViewModel.Post.Title;
